+++ lib/flutter_3d_carousel.dart
@@ -1,6 +1,10 @@
 library;
 
 export 'src/components/carousel_widget_3d.dart' show CarouselWidget3D;
-export 'src/models/drag_behaviour.dart' show DragEndBehavior;
-export 'src/models/tap_behavior.dart' show TapBehavior;
+export 'src/models/models.dart'
+    show
+        CarouselChild,
+        BackgroundTapBehavior,
+        ChildTapBehavior,
+        DragEndBehavior;
 export 'src/utils/util_functions.dart' show UtilFunctions;