+++ lib/src/models/carousel_child.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+@immutable
+class CarouselChild {
+  final Widget child;
+  final VoidCallback onTap;
+
+  const CarouselChild({
+    required this.child,
+    required this.onTap,
+  });
+}