+++ CHANGELOG.md
@@ -1,13 +1,23 @@
 
+## Version 1.0.3 - 
+
+### Changes:
+- Added widget property `background`.
+- Added bool property `onlyRenderForeground`,
+- Changed `TapBehavior` to `BackgroundTapBehavior` and `tapBehavior` to `backgroundTapBehavior`.
+- Added enum `ChildTapBehaviour` and `childTapBehaviour` property.
+- Fixed bug with carousel sizing.
+
+
 ## Version 1.0.2 - 2025-05-13
 
-### Added
+### Changes:
 - Replaced asset screen recording gif in readme.
 
 
 ## Version 1.0.1 - 2025-05-13
 
-### Added
+### Changes:
 - Improved description in pubspec.yaml.
 - Minor formatting fix.
 - Added code documentation to UtilFunctions
@@ -19,4 +29,3 @@
 ### Added
 - Core widgets and functionality: CarouselWidget3D.
 - Example usage
-